+++ CHANGELOG.md
@@ -36,6 +36,7 @@
 - Fixed: Issue where File Converter version upgrade download was not working due to an issue with https encryption.
 - Fixed: Issue where output video was not working correclty on some video players like Quick time (github issue #34) (thanks to Diego López Bugna).
 - Fixed: Issue where icons and images were blurry on high dpi device.
+- Fixed: Issue where file 'Error' in there name were generating false negative result (github issue #247).
 - Tech: Complete rework of the project architecture to be able to improve it in a long term perspective. The project is now using the MVVM Community Toolkit and is following more closely this design pattern.
 - Tech: Update ffmpeg to v6.1.1 and ImageMagick to v13.5.
 - Tech: Update SharpShell to v2.7.2.